requirements.txt
numpy
Pillow
opencv-python
tkinter
